renamed the c++ headers file from .h to .hpp
