+++ include/BaseMathFunctions.hpp
