+++ include/eventsHandler.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "./gameClasses.h"
+#include "./gameClasses.hpp"
 #include<SFML/Graphics.hpp>
 // function used to handle events like a mouse button press
 int handleEvent(sf::Event, Plane*, bool*);
